polishing off chern_insulators examples
