+++ examples/chern_insulators/haldane_C3.py
@@ -9,48 +9,8 @@ import numpy as np
 
 from tenpy.algorithms import dmrg
 from tenpy.networks.mps import MPS
-
-# TODO: move model in tenpy/models/
-from tenpy.models.model import CouplingMPOModel
-from tenpy.tools.params import get_parameter
-from tenpy.networks.site import FermionSite, GroupedSite
-
-from tenpy.models import lattice
+from tenpy.models.fermions_haldane_C3 import FermionicC3HaldaneModel, TripartiteTriangular
 from tenpy.networks import site
-import sys
-
-
-class TripartiteTriangular(lattice.Lattice):
-
-    def __init__(self, Lx, Ly, siteA, **kwargs):
-
-        basis = np.array(([3., 0.], [0.5, 0.5*np.sqrt(3)]))
-
-        pos = np.array(([0., 0.], [1., 0.], [2., 0.]))
-
-        kwargs.setdefault('order', 'default')
-        kwargs.setdefault('bc', 'periodic')
-        kwargs.setdefault('bc_MPS', 'infinite')
-        kwargs.setdefault('basis', basis)
-        kwargs.setdefault('positions', pos)
-
-        super().__init__([Lx, Ly], [siteA, siteA, siteA], **kwargs)
-
-        self.NN = [(0, 2, np.array([-1, 1])), (0, 1, np.array([0, 0])), (0, 0, np.array([0, -1])),
-                   (1, 0, np.array([0, 1])), (1, 2, np.array([0, 0])), (1, 1, np.array([0, -1])),
-                   (2, 1, np.array([0, 1])), (2, 0, np.array([1, 0])), (2, 2, np.array([0, -1]))]
-
-        self.nNNA = [(0, 2, np.array([-1, 2])), (0, 2, np.array([0, -1])), (0, 2, np.array([-1, -1])),
-                     (1, 0, np.array([0, 2])), (1, 0, np.array([1, -1])), (1, 0, np.array([0, -1])),
-                     (2, 1, np.array([0, 2])), (2, 1, np.array([1, -1])), (2, 1, np.array([0, -1]))]
-
-        self.nNNB = [(0, 1, np.array([0, 1])), (0, 1, np.array([-1, 1])), (0, 1, np.array([0, -2])),
-                     (1, 2, np.array([0, 1])), (1, 2, np.array([-1, 1])), (1, 2, np.array([0, -2])),
-                     (2, 0, np.array([1, 1])), (2, 0, np.array([0, 1])), (2, 0, np.array([1, -2]))]
-
-        self.nnNN = [(0, 1, np.array([-1, 2])), (0, 2, np.array([0, 0])), (0, 0, np.array([0, -2])),
-                     (1, 2, np.array([-1, 2])), (1, 0, np.array([1, 0])), (1, 1, np.array([0, -2])),
-                     (2, 0, np.array([0, 2])), (2, 1, np.array([1, 0])), (2, 2, np.array([0, -2]))]
 
 
 def plot_lattice():
@@ -68,82 +28,7 @@ def plot_lattice():
     plt.show()
 
 
-class FermionicC3HaldaneModel(CouplingMPOModel):
-
-    def __init__(self, model_params):
-
-        CouplingMPOModel.__init__(self, model_params)
-
-    def init_sites(self, model_params):
-
-        conserve = get_parameter(model_params, 'conserve', 'N', self.name)
-        fs = FermionSite(conserve=conserve)
-
-        gs = GroupedSite([fs, fs], labels=['A', 'B'], charges='same')
-        gs.add_op('Ntot', gs.NA + gs.NB, False)
-
-        return gs
-
-    def init_lattice(self, model_params):
-
-        choice = get_parameter(model_params, 'lattice', 'TripartiteTriangular', self.name)
-
-        if choice != 'TripartiteTriangular':
-            sys.exit("Error: Please choose the TripartiteTriangular for C3_haldane.")
-
-        Lx = get_parameter(model_params, 'Lx', 1, self.name)
-        Ly = get_parameter(model_params, 'Ly', 3, self.name)
-
-        fs = self.init_sites(model_params)
-
-        lat = TripartiteTriangular(Lx, Ly, fs)
-
-        print(lat.N_sites)
-
-        return lat
-
-    def init_terms(self, model_params):
-
-        t = get_parameter(model_params, 't', -1., self.name, True)
-        V = get_parameter(model_params, 'V', 0, self.name, True)
-        mu = get_parameter(model_params, 'mu', 0., self.name, True)
-        phi_ext = 2*np.pi*get_parameter(model_params, 'phi_ext', 0., self.name)
-
-        t1 = t
-        t2 = 0.39*t*1j
-        t3 = -0.34*t
-
-        for u in range(len(self.lat.unit_cell)):
-
-            self.add_onsite(mu, 0, 'N', category='mu N')
-            self.add_onsite(-mu, 1, 'N', category='mu N')
-
-        for u1, u2, dx in self.lat.NN:
-
-            t1_phi = self.coupling_strength_add_ext_flux(t1, dx, [0, phi_ext])
-            self.add_coupling(t1_phi, u1, 'CdA', u2, 'CB', dx, 'JW', True)
-            self.add_coupling(np.conj(t1_phi), u2, 'CdB', u1, 'CA', -dx, 'JW', True)
-
-        for u1, u2, dx in self.lat.nNNA:
-
-            t2_phi = self.coupling_strength_add_ext_flux(t2, dx, [0, phi_ext])
-            self.add_coupling(t2_phi, u1, 'CdA', u2, 'CA', dx, 'JW', True)
-            self.add_coupling(np.conj(t2_phi), u2, 'CdA', u1, 'CA', -dx, 'JW', True)
-
-        for u1, u2, dx in self.lat.nNNB:
-
-            t2_phi = self.coupling_strength_add_ext_flux(t2, dx, [0, phi_ext])
-            self.add_coupling(t2_phi, u1, 'CdB', u2, 'CB', dx, 'JW', True)
-            self.add_coupling(np.conj(t2_phi), u2, 'CdB', u1, 'CB', -dx, 'JW', True)
-
-        for u1, u2, dx in self.lat.nnNN:
-
-            t3_phi = self.coupling_strength_add_ext_flux(t3, dx, [0, phi_ext])
-            self.add_coupling(t3_phi, u1, 'CdA', u2, 'CB', dx, 'JW', True)
-            self.add_coupling(np.conj(t3_phi), u2, 'CdB', u1, 'CA', -dx, 'JW', True)
-
-
-def run(phi_ext=np.linspace(0, 1.0, 5)):
+def run(phi_ext=np.linspace(0, 1.0, 7)):
 
     data = dict(phi_ext=phi_ext, QL=[], ent_spectrum=[])
 
@@ -163,7 +48,7 @@ def run(phi_ext=np.linspace(0, 1.0, 5)):
             'N_min': 5,
             'N_max': 20
         },
-        'chi_list': {0: 9, 10: 49, 20: 200},
+        'chi_list': {0: 9, 10: 49, 20: 100},
         'max_E_err': 1.e-10,
         'max_S_err': 1.e-6,
         'max_sweeps': 150,
@@ -171,8 +56,6 @@ def run(phi_ext=np.linspace(0, 1.0, 5)):
     }
 
     prod_state = ['full_A empty_B', 'empty_A full_B', 'full_A empty_B'] * (model_params['Lx'] * model_params['Ly'])
-
-    print(prod_state)
 
     eng = None
 
@@ -185,9 +68,6 @@ def run(phi_ext=np.linspace(0, 1.0, 5)):
 
         if eng is None:  # first time in the loop
             M = FermionicC3HaldaneModel(model_params)
-
-            print("sites = ", M.lat.mps_sites())
-
             psi = MPS.from_product_state(M.lat.mps_sites(), prod_state, bc=M.lat.bc_MPS)
             eng = dmrg.EngineCombine(psi, M, dmrg_params)
         else:
@@ -211,8 +91,8 @@ def plot_results(data):
     ax = plt.gca()
     ax.plot(data['phi_ext'], data['QL'], marker='o')
     ax.set_xlabel(r"$\Phi_y / 2 \pi$")
-    ax.set_ylabel(r"$ \langle Q^L \rangle$")
-    plt.show()
+    ax.set_ylabel(r"$ \langle Q^L(\Phi_y) \rangle$")
+    plt.savefig("haldane_C3_charge_pump.pdf")
 
     plt.figure()
     ax = plt.gca()
@@ -231,7 +111,7 @@ def plot_results(data):
     ax.set_ylabel(r"$ \epsilon_\alpha $")
     ax.set_ylim(0., 8.)
     ax.legend(loc='upper right')
-    plt.show()
+    plt.savefig("haldane_C3_ent_spec_flow.pdf")
 
 
 if __name__ == "__main__":
